+++ components/Marker/utils.ts
@@ -1,5 +1,4 @@
 import { Icon } from 'leaflet';
-
 import { EPinType } from '~/lib/types';
 
 export const getIcon = (type: EPinType): Icon =>
@@ -7,3 +6,12 @@ export const getIcon = (type: EPinType): Icon =>
     iconUrl: `/images/markers/${type}.png`,
     iconSize: [45, 90]
   });
+
+export const getFullDateString = (date: string) => {
+  return new Date(date).toLocaleDateString(undefined, {
+    weekday: 'long',
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric'
+  });
+};
